refactor package CRUD module and refresh review status notes
