+++ proj/260210-stack-integrated/review/00-SUMMARY.md
@@ -1,5 +1,10 @@
 # Code Quality Review Summary
 
+> **Status Update (2026-02-17, `pr-review-fix`):** The maintainability blocker from this snapshot is still partially remediated.
+> Import-cycle remediation and validation gates are green (`scripts/detect_cycles.py` reports no cycles, `mypy`/`make test-all`/remote docker target pass).
+> `tools/packages.py` was split and reduced below 1000 LOC (758), and the production TODO in that module was removed.
+> Large-module structural debt still remains in other files (>1000 LOC).
+
 **Date:** 2026-02-17
 **Reviewer:** Codex
 **Commit:** c592323
