+++ proj/260210-stack-integrated/review/03-maintainability.md
@@ -1,5 +1,9 @@
 # 03 - Code Maintainability
 
+> **Status Note (2026-02-17):** This document is a historical snapshot from 2026-02-16 and is now partially outdated.
+> Current deltas from this branch: `src/quilt_mcp/tools/packages.py` reduced from 2034 to 758 LOC via extraction to `src/quilt_mcp/tools/package_crud.py`, `scripts/detect_cycles.py` reports no import cycles, and no `TODO/FIXME` markers remain under `src/quilt_mcp/{backends,tools,ops}`.
+> Remaining maintainability debt: multiple modules still exceed 1000 LOC (`ops/quilt_ops.py`, `services/governance_service.py`, `backends/platform_admin_ops.py`, `backends/platform_backend.py`, `services/workflow_service.py`, `tools/responses.py`).
+
 **Date:** 2026-02-16  
 **Reviewer:** Codex
