+++ src/quilt_mcp/tools/packages.py
@@ -281,8 +281,8 @@ def _create_enhanced_package(
         top_hash = result.top_hash
         logger.info(f"Successfully created package {package_name} with hash {top_hash}")
 
-        # TODO: Handle summary files and visualizations in future enhancement
-        # For now, basic package creation with README is supported
+        # NOTE: Summary files/visualization artifact attachment is intentionally deferred in this path.
+        # Basic package creation with README content is supported.
         if summary_files:
             logger.warning("Summary files and visualizations not yet supported with create_package_revision")
 
@@ -297,1001 +297,15 @@ def _create_enhanced_package(
         raise
 
 
-def packages_list(
-    registry: Annotated[
-        str,
-        Field(
-            default="",
-            description="Optional Quilt registry S3 URI to list packages from. Empty string lists all accessible packages.",
-        ),
-    ] = "",
-    limit: Annotated[
-        int,
-        Field(
-            default=0,
-            ge=0,
-            description="Maximum number of packages to return, 0 for unlimited",
-        ),
-    ] = 0,
-    prefix: Annotated[
-        str,
-        Field(
-            default="",
-            description="Filter packages by name prefix",
-            examples=["", "team/", "user/analysis-"],
-        ),
-    ] = "",
-) -> PackagesListSuccess | PackagesListError:
-    """List all available Quilt packages in a registry - Quilt package discovery and comparison tasks
-
-    Args:
-        registry: Optional Quilt registry S3 URI to list packages from. Empty string lists all accessible packages.
-        limit: Maximum number of packages to return, 0 for unlimited
-        prefix: Filter packages by name prefix
-
-    Returns:
-        PackagesListSuccess with list of package names, or PackagesListError on failure.
-
-    Next step:
-        Surface package details to the user or feed identifiers into downstream package tools.
-
-    Example:
-        ```python
-        from quilt_mcp.tools import packages
-
-        result = packages.packages_list(registry="s3://my-bucket", limit=10)
-        # Next step: Surface package details to the user or feed identifiers into downstream package tools.
-        ```
-    """
-    try:
-        # If no registry specified, could implement catalog-wide listing in future
-        # For now, require registry parameter
-        if not registry:
-            return PackagesListError(
-                error="Registry parameter is required for package listing. Specify target S3 bucket (e.g., registry='s3://my-bucket')",
-                registry="",
-                suggested_actions=[
-                    "Provide a registry parameter",
-                    "Example: packages_list(registry='s3://my-bucket')",
-                    "Use bucket_recommendations_get() to find accessible buckets",
-                ],
-            )
-
-        # Normalize registry and use QuiltOps.search_packages() for listing
-        normalized_registry = _normalize_registry(registry)
-        # Suppress stdout during search_packages to avoid JSON-RPC interference
-        from ..utils.common import suppress_stdout
-        from ..ops.factory import QuiltOpsFactory
-
-        quilt_ops = QuiltOpsFactory.create()
-        with suppress_stdout():
-            # Use empty query to list all packages, then extract names from Package_Info objects
-            package_infos = quilt_ops.search_packages(query="", registry=normalized_registry)
-            pkgs = [pkg_info.name for pkg_info in package_infos]
-
-        # Apply prefix filtering if specified
-        if prefix:
-            pkgs = [pkg for pkg in pkgs if pkg.startswith(prefix)]
-
-        # Apply limit if specified
-        if limit > 0:
-            pkgs = pkgs[:limit]
-
-        return PackagesListSuccess(
-            packages=pkgs,
-            count=len(pkgs),
-            registry=registry,
-            prefix_filter=prefix if prefix else None,
-        )
-    except Exception as e:
-        return PackagesListError(
-            error=f"Failed to list packages: {str(e)}",
-            registry=registry,
-            suggested_actions=[
-                "Verify registry is accessible",
-                "Check AWS permissions for the bucket",
-                "Ensure registry S3 URI is valid (e.g., s3://my-bucket)",
-                "Use bucket_recommendations_get() to find accessible buckets",
-            ],
-        )
-
-
-def package_browse(
-    package_name: Annotated[
-        str,
-        Field(
-            description="Name of the package in namespace/name format",
-            examples=["username/dataset", "team/analysis-results"],
-            pattern=r"^[a-zA-Z0-9_-]+/[a-zA-Z0-9_-]+$",
-        ),
-    ],
-    registry: Annotated[
-        str,
-        Field(
-            description="Quilt registry S3 URI (REQUIRED)",
-            examples=["s3://my-bucket", "s3://quilt-example"],
-        ),
-    ],
-    recursive: Annotated[
-        bool,
-        Field(
-            default=True,
-            description="Show full file tree (true) or just top-level entries (false)",
-        ),
-    ] = True,
-    include_file_info: Annotated[
-        bool,
-        Field(
-            default=True,
-            description="Include file sizes, types, and modification dates",
-        ),
-    ] = True,
-    max_depth: Annotated[
-        int,
-        Field(
-            default=0,
-            ge=0,
-            description="Maximum directory depth to show (0 for unlimited)",
-        ),
-    ] = 0,
-    top: Annotated[
-        int,
-        Field(
-            default=0,
-            ge=0,
-            description="Limit number of entries returned (0 for unlimited)",
-        ),
-    ] = 0,
-    include_signed_urls: Annotated[
-        bool,
-        Field(
-            default=True,
-            description="Include presigned download URLs for S3 objects",
-        ),
-    ] = True,
-) -> PackageBrowseSuccess | ErrorResponse:
-    """Browse the contents of a Quilt package with enhanced file information - Quilt package discovery and comparison tasks
-
-    Args:
-        package_name: Name of the package in namespace/name format
-        registry: Quilt registry S3 URI (REQUIRED)
-        recursive: Show full file tree (true) or just top-level entries (false)
-        include_file_info: Include file sizes, types, and modification dates
-        max_depth: Maximum directory depth to show (0 for unlimited)
-        top: Limit number of entries returned (0 for unlimited)
-        include_signed_urls: Include presigned download URLs for S3 objects
-
-    Returns:
-        PackageBrowseSuccess with comprehensive package contents, or ErrorResponse on failure.
-
-    Examples:
-        Basic browsing:
-        package_browse(package_name="team/dataset", registry="s3://my-bucket")
-
-        Flat view (top-level only):
-        package_browse(package_name="team/dataset", registry="s3://my-bucket", recursive=False)
-
-        Limited depth:
-        package_browse(package_name="team/dataset", registry="s3://my-bucket", max_depth=2)
-
-    Next step:
-        Surface package details to the user or feed identifiers into downstream package tools.
-
-    Example:
-        ```python
-        from quilt_mcp.tools import packages
-
-        result = packages.package_browse(package_name="team/dataset", registry="s3://my-bucket")
-        # Next step: Surface package details to the user or feed identifiers into downstream package tools.
-        ```
-    """
-    # Validate required registry parameter
-    ok_registry, registry_error, registry_actions = validate_registry_required(registry, "package_browse")
-    if not ok_registry:
-        return ErrorResponse(
-            error=registry_error,
-            possible_fixes=registry_actions,
-            suggested_actions=[
-                "Use bucket_recommendations_get() to find accessible buckets",
-                "Check which bucket contains your package",
-            ],
-        )
-
-    # Use the provided registry
-    normalized_registry = _normalize_registry(registry)
-    try:
-        # Suppress stdout during browse to avoid JSON-RPC interference
-        from ..utils.common import suppress_stdout
-        from ..ops.factory import QuiltOpsFactory
-
-        quilt_ops = QuiltOpsFactory.create()
-        with suppress_stdout():
-            # Use QuiltOps.browse_content() to get Content_Info objects directly
-            content_infos = quilt_ops.browse_content(package_name, registry=normalized_registry, path="")
-
-            # Also get package metadata through public QuiltOps API
-            try:
-                pkg_metadata = quilt_ops.get_package_metadata(package_name, registry=normalized_registry)
-            except Exception as meta_error:
-                logger.warning(f"Could not retrieve package metadata: {meta_error}")
-                pkg_metadata = None
-
-    except Exception as e:
-        return ErrorResponse(
-            error=f"Failed to browse package '{package_name}'",
-            cause=str(e),
-            possible_fixes=[
-                "Verify the package name is correct",
-                "Check if you have access to the registry",
-                "Ensure the package exists in the specified registry",
-            ],
-            suggested_actions=[
-                f"Try: packages_list(registry='{registry}') to see available packages",
-                f"Try: search_catalog(query='{package_name.split('/')[-1]}', scope='package') to find similar packages",
-            ],
-        )
-
-    # Process Content_Info objects directly instead of using MockPackage
-    entries = []
-    file_tree: dict[str, Any] | None = {} if recursive else None
-    total_size = 0
-    file_types = set()
-
-    # Apply top limit if specified
-    content_list = content_infos[:top] if top > 0 else content_infos
-
-    for content_info in content_list:
-        try:
-            logical_key = content_info.path
-            file_size = content_info.size
-            is_directory = content_info.type == "directory"
-
-            # Determine file type and properties
-            file_ext = logical_key.split(".")[-1].lower() if "." in logical_key and not is_directory else "unknown"
-            file_types.add(file_ext)
-
-            # Track total size
-            if file_size:
-                total_size += file_size
-
-            entry_data = {
-                "logical_key": logical_key,
-                "physical_key": None,  # Will be set below if available
-                "size": file_size,
-                "size_human": _format_file_size(file_size) if file_size else None,
-                "hash": "unknown",  # Content_Info doesn't provide hash
-                "file_type": file_ext,
-                "is_directory": is_directory,
-            }
-
-            # Add enhanced file info if requested and we have a download URL
-            if include_file_info and content_info.download_url:
-                entry_data["physical_key"] = content_info.download_url
-                if content_info.modified_date:
-                    entry_data["last_modified"] = content_info.modified_date
-
-            # Add download URL if available and requested
-            if include_signed_urls and content_info.download_url:
-                entry_data["download_url"] = content_info.download_url
-                entry_data["s3_uri"] = content_info.download_url
-
-            entries.append(entry_data)
-
-            # Build file tree structure for recursive view
-            if recursive and file_tree is not None:
-                _add_to_file_tree(file_tree, logical_key, entry_data, max_depth)
-
-        except Exception as e:
-            # Include entry with error info
-            entries.append(
-                {
-                    "logical_key": content_info.path if hasattr(content_info, 'path') else "unknown",
-                    "physical_key": None,
-                    "size": None,
-                    "hash": "",
-                    "error": str(e),
-                    "file_type": "error",
-                }
-            )
-
-    # Prepare comprehensive response
-    summary = PackageSummary(
-        total_size=total_size,
-        total_size_human=_format_file_size(total_size),
-        file_types=sorted(list(file_types)),
-        total_files=len([e for e in entries if not e.get("is_directory", False)]),
-        total_directories=len([e for e in entries if e.get("is_directory", False)]),
-    )
-
-    return PackageBrowseSuccess(
-        package_name=package_name,
-        registry=registry,
-        total_entries=len(entries),
-        summary=summary,
-        view_type="recursive" if recursive else "flat",
-        file_tree=file_tree if recursive and file_tree else None,
-        entries=entries,
-        metadata=pkg_metadata,
-    )
-
-
-def _add_to_file_tree(tree: dict, path: str, entry_data: dict, max_depth: int):
-    """Add an entry to the file tree structure."""
-    if max_depth > 0:
-        depth = path.count("/")
-        if depth >= max_depth:
-            return
-
-    parts = path.split("/")
-    current = tree
-
-    # Navigate to the correct position in the tree
-    for i, part in enumerate(parts[:-1]):
-        if part not in current:
-            current[part] = {"type": "directory", "children": {}}
-        current = current[part]["children"]
-
-    # Add the final entry
-    final_part = parts[-1]
-    current[final_part] = {
-        "type": "file" if not entry_data.get("is_directory") else "directory",
-        "size": entry_data.get("size"),
-        "size_human": entry_data.get("size_human"),
-        "file_type": entry_data.get("file_type"),
-        "physical_key": entry_data.get("physical_key"),
-        "download_url": entry_data.get("download_url"),
-    }
-
-
-def _format_file_size(size_bytes: int) -> str:
-    """Format file size in human-readable format."""
-    if size_bytes is None:
-        return "Unknown"
-
-    size_float = float(size_bytes)
-    for unit in ["B", "KB", "MB", "GB", "TB"]:
-        if size_float < 1024.0:
-            return f"{size_float:.1f} {unit}"
-        size_float /= 1024.0
-    return f"{size_float:.1f} PB"
-
-
-def package_diff(
-    package1_name: Annotated[
-        str,
-        Field(
-            description="Name of the first package in namespace/name format",
-            examples=["username/dataset", "team/analysis-v1"],
-            pattern=r"^[a-zA-Z0-9_-]+/[a-zA-Z0-9_-]+$",
-        ),
-    ],
-    package2_name: Annotated[
-        str,
-        Field(
-            description="Name of the second package in namespace/name format",
-            examples=["username/dataset", "team/analysis-v2"],
-            pattern=r"^[a-zA-Z0-9_-]+/[a-zA-Z0-9_-]+$",
-        ),
-    ],
-    registry: Annotated[
-        str,
-        Field(
-            description="Quilt registry S3 URI (REQUIRED)",
-        ),
-    ],
-    package1_hash: Annotated[
-        str,
-        Field(
-            default="",
-            description="Optional specific hash for first package (empty string for latest)",
-        ),
-    ] = "",
-    package2_hash: Annotated[
-        str,
-        Field(
-            default="",
-            description="Optional specific hash for second package (empty string for latest)",
-        ),
-    ] = "",
-) -> PackageDiffSuccess | PackageDiffError:
-    """Compare two package versions and show differences - Quilt package discovery and comparison tasks
-
-    Args:
-        package1_name: Name of the first package in namespace/name format
-        package2_name: Name of the second package in namespace/name format
-        registry: Quilt registry S3 URI (REQUIRED)
-        package1_hash: Optional specific hash for first package (empty string for latest)
-        package2_hash: Optional specific hash for second package (empty string for latest)
-
-    Returns:
-        PackageDiffSuccess with differences, or PackageDiffError on failure.
-
-    Next step:
-        Surface package details to the user or feed identifiers into downstream package tools.
-
-    Example:
-        ```python
-        from quilt_mcp.tools import packages
-
-        result = packages.package_diff(
-            package1_name="team/dataset-v1",
-            package2_name="team/dataset-v2",
-            registry="s3://my-bucket",
-        )
-        # Next step: Surface package details to the user or feed identifiers into downstream package tools.
-        ```
-    """
-    # Validate required registry parameter
-    if not registry:
-        return PackageDiffError(
-            error="Registry parameter is required for package diff. Specify target S3 bucket (e.g., registry='s3://my-bucket')",
-            package1=package1_name,
-            package2=package2_name,
-            suggested_actions=[
-                "Provide registry parameter",
-                "Example: package_diff(package1_name='team/v1', package2_name='team/v2', registry='s3://my-bucket')",
-            ],
-        )
-
-    normalized_registry = _normalize_registry(registry)
-
-    try:
-        # Use QuiltOps.diff_packages() for business logic
-        from ..ops.factory import QuiltOpsFactory
-        from ..utils.common import suppress_stdout
-
-        quilt_ops = QuiltOpsFactory.create()
-        with suppress_stdout():
-            diff_dict = quilt_ops.diff_packages(
-                package1_name=package1_name,
-                package2_name=package2_name,
-                registry=normalized_registry,
-                package1_hash=package1_hash if package1_hash else None,
-                package2_hash=package2_hash if package2_hash else None,
-            )
-
-        return PackageDiffSuccess(
-            package1=package1_name,
-            package2=package2_name,
-            package1_hash=package1_hash if package1_hash else "latest",
-            package2_hash=package2_hash if package2_hash else "latest",
-            registry=registry,
-            diff=diff_dict,
-        )
-
-    except Exception as e:
-        return PackageDiffError(
-            error=f"Failed to diff packages: {e}",
-            package1=package1_name,
-            package2=package2_name,
-        )
-
-
-def package_create(
-    package_name: Annotated[
-        str,
-        Field(
-            description="Name for the new package in namespace/name format",
-            examples=["username/dataset", "team/analysis-results"],
-            pattern=r"^[a-zA-Z0-9_-]+/[a-zA-Z0-9_-]+$",
-        ),
-    ],
-    s3_uris: Annotated[
-        list[str],
-        Field(
-            description="List of S3 URIs to include in the package",
-            examples=[["s3://bucket/file1.csv", "s3://bucket/file2.json"]],
-            min_length=1,
-        ),
-    ],
-    registry: Annotated[
-        str,
-        Field(
-            description="Target Quilt registry S3 URI (REQUIRED)",
-        ),
-    ],
-    metadata: Annotated[
-        Optional[dict[str, Any]],
-        Field(
-            default=None,
-            description="Optional metadata to attach to the package (JSON object)",
-            examples=[{"description": "My dataset", "version": "1.0"}],
-        ),
-    ] = None,
-    message: Annotated[
-        str,
-        Field(
-            default="Created via package_create tool",
-            description="Commit message for package creation",
-        ),
-    ] = "Created via package_create tool",
-    flatten: Annotated[
-        bool,
-        Field(
-            default=True,
-            description="Use only filenames as logical paths (true) instead of full S3 keys (false)",
-        ),
-    ] = True,
-    copy: Annotated[
-        bool,
-        Field(
-            default=False,
-            description="Whether to copy files to registry bucket (false=reference only, true=copy all)",
-        ),
-    ] = False,
-) -> PackageCreateSuccess | PackageCreateError:
-    """Create a new Quilt package from S3 objects - Core package creation, update, and deletion workflows
-
-    Args:
-        package_name: Name for the new package in namespace/name format
-        s3_uris: List of S3 URIs to include in the package
-        registry: Target Quilt registry S3 URI (REQUIRED)
-        metadata: Optional metadata to attach to the package (JSON object)
-        message: Commit message for package creation
-        flatten: Use only filenames as logical paths (true) instead of full S3 keys (false)
-        copy: Whether to copy files to registry bucket
-
-    Returns:
-        PackageCreateSuccess with package details, or PackageCreateError on failure.
-
-    Examples:
-        Basic package creation:
-        package_create(package_name="my-team/dataset", s3_uris=["s3://bucket/file.csv"], registry="s3://my-bucket")
-
-        With metadata:
-        package_create(
-            package_name="my-team/dataset",
-            s3_uris=["s3://bucket/file.csv"],
-            registry="s3://my-bucket",
-            metadata={"description": "My dataset", "type": "research"}
-        )
-
-    Next step:
-        Report the package operation result or continue the workflow (e.g., metadata updates).
-
-    Example:
-        ```python
-        from quilt_mcp.tools import packages
-
-        result = packages.package_create(
-            package_name="team/dataset",
-            s3_uris=["s3://example-bucket/data.csv"],
-            registry="s3://my-bucket",
-        )
-        # Next step: Report the package operation result or continue the workflow (e.g., metadata updates).
-        ```
-    """
-    # Initialize auth_ctx to avoid fragile locals() checks in exception handlers
-    auth_ctx: AuthorizationContext | None = None
-
-    # Validate required registry parameter
-    ok_registry, registry_error, registry_actions = validate_registry_required(registry, "package_create")
-    if not ok_registry:
-        return PackageCreateError(
-            error=registry_error,
-            package_name=package_name,
-            registry="",
-            suggested_actions=registry_actions + ["Use bucket_recommendations_get() to find writable buckets"],
-        )
-
-    warnings: list[str] = []
-    ok_uris, uris_error, uris_actions = validate_s3_uris_required(s3_uris)
-    if not ok_uris:
-        return PackageCreateError(
-            error=uris_error,
-            package_name=package_name,
-            suggested_actions=uris_actions,
-        )
-    ok_package, package_error, package_actions = validate_package_name_required(package_name, "package_create")
-    if not ok_package:
-        return PackageCreateError(
-            error=package_error,
-            package_name="",
-            suggested_actions=package_actions,
-        )
-    ok_meta, meta_error, _, meta_actions = validate_metadata_dict(metadata, package_name=package_name)
-    if not ok_meta:
-        return PackageCreateError(
-            error=meta_error,
-            package_name=package_name,
-            provided_type=type(metadata).__name__,
-            suggested_actions=meta_actions,
-        )
-
-    # Process metadata to remove README fields (not currently preserved)
-    processed_metadata = metadata.copy() if metadata else {}
-    processed_metadata.pop("readme_content", None)
-    processed_metadata.pop("readme", None)
-
-    normalized_registry = _normalize_registry(registry)
-
-    auth_ctx, error = _authorize_package(
-        "package_create",
-        {"package_name": package_name, "registry": normalized_registry},
-        context={"package_name": package_name, "registry": normalized_registry},
-    )
-    if error:
-        return PackageCreateError(
-            error=error.get("error", "Authorization failed"),
-            package_name=package_name,
-            registry=registry,
-            suggested_actions=["Check your permissions", "Verify package name format", "Confirm registry access"],
-        )
-
-    try:
-        # Use QuiltOps.create_package_revision method with auto_organize=False
-        # to preserve the existing flattening behavior
-        quilt_ops = QuiltOpsFactory.create()
-        result = quilt_ops.create_package_revision(
-            package_name=package_name,
-            s3_uris=s3_uris,
-            metadata=processed_metadata,
-            registry=normalized_registry,
-            message=message,
-            auto_organize=False,  # Preserve flattening behavior like _collect_objects_into_package
-            copy=copy,
-        )
-
-        # Handle the result - it's now a Package_Creation_Result domain object
-        if not result.success:
-            return PackageCreateError(
-                error="Package creation failed",
-                package_name=package_name,
-                registry=registry,
-                suggested_actions=[
-                    "Verify S3 URIs are valid and accessible",
-                    "Check write permissions for the registry",
-                    "Ensure no duplicate logical paths exist",
-                ],
-                warnings=warnings,
-            )
-
-        # Extract details from the Package_Creation_Result domain object
-        top_hash = result.top_hash
-        entries_added = result.file_count
-        files: list[dict[str, Any]] = []  # Files list not available in domain object yet
-
-        # Build package URL
-        from .catalog import catalog_url
-
-        catalog_result = catalog_url(
-            registry=normalized_registry,
-            package_name=package_name,
-            path="",
-            catalog_host="",
-        )
-        package_url = catalog_result.catalog_url if isinstance(catalog_result, CatalogUrlSuccess) else ""
-
-        return PackageCreateSuccess(
-            package_name=package_name,
-            registry=registry,
-            top_hash=str(top_hash),
-            files_added=entries_added,
-            package_url=package_url,
-            files=files,
-            message=message,
-            warnings=warnings,
-            auth_type=auth_ctx.auth_type if auth_ctx else None,
-        )
-
-    except Exception as e:
-        return PackageCreateError(
-            error=f"Failed to create package: {e}",
-            package_name=package_name,
-            registry=registry,
-            suggested_actions=[
-                "Check S3 permissions for source files",
-                "Verify registry write access",
-                "Ensure package name is valid",
-                "Check network connectivity",
-            ],
-            warnings=warnings,
-        )
-
-
-def package_update(
-    package_name: Annotated[
-        str,
-        Field(
-            description="Name of the existing package to update in namespace/name format",
-            examples=["username/dataset", "team/analysis-results"],
-            pattern=r"^[a-zA-Z0-9_-]+/[a-zA-Z0-9_-]+$",
-        ),
-    ],
-    s3_uris: Annotated[
-        list[str],
-        Field(
-            description="List of S3 URIs to add to the package",
-            examples=[["s3://bucket/newfile.csv", "s3://bucket/updated.json"]],
-            min_length=1,
-        ),
-    ],
-    registry: Annotated[
-        str,
-        Field(
-            description="Target Quilt registry S3 URI (REQUIRED)",
-        ),
-    ],
-    metadata: Annotated[
-        Optional[dict[str, Any]],
-        Field(
-            default=None,
-            description="Optional metadata to merge with existing package metadata",
-            examples=[{"updated": "true", "version": "2.0"}],
-        ),
-    ] = None,
-    message: Annotated[
-        str,
-        Field(
-            default="Added objects via package_update tool",
-            description="Commit message for package update",
-        ),
-    ] = "Added objects via package_update tool",
-    flatten: Annotated[
-        bool,
-        Field(
-            default=True,
-            description="Use only filenames as logical paths (true) instead of full S3 keys (false)",
-        ),
-    ] = True,
-    copy: Annotated[
-        bool,
-        Field(
-            default=False,
-            description="Whether to copy files to registry bucket (false=reference only, true=copy all)",
-        ),
-    ] = False,
-) -> PackageUpdateSuccess | PackageUpdateError:
-    """Update an existing Quilt package by adding new S3 objects - Core package creation, update, and deletion workflows
-
-    Args:
-        package_name: Name of the existing package to update in namespace/name format
-        s3_uris: List of S3 URIs to add to the package
-        registry: Target Quilt registry S3 URI (REQUIRED)
-        metadata: Optional metadata to merge with existing package metadata
-        message: Commit message for package update
-        flatten: Use only filenames as logical paths (true) instead of full S3 keys (false)
-        copy: Whether to copy files to registry bucket
-
-    Returns:
-        PackageUpdateSuccess with update details, or PackageUpdateError on failure.
-
-    Next step:
-        Report the package operation result or continue the workflow (e.g., metadata updates).
-
-    Example:
-        ```python
-        from quilt_mcp.tools import packages
-
-        result = packages.package_update(
-            package_name="team/dataset",
-            s3_uris=["s3://example-bucket/data.csv"],
-            registry="s3://my-bucket",
-        )
-        # Next step: Report the package operation result or continue the workflow (e.g., metadata updates).
-        ```
-    """
-    # Initialize auth_ctx to avoid fragile locals() checks in exception handlers
-    auth_ctx: AuthorizationContext | None = None
-
-    # Validate required registry parameter
-    ok_registry, registry_error, registry_actions = validate_registry_required(registry, "package_update")
-    if not ok_registry:
-        return PackageUpdateError(
-            error=registry_error,
-            package_name=package_name,
-            registry="",
-            suggested_actions=registry_actions,
-        )
-
-    ok_uris, uris_error, uris_actions = validate_s3_uris_required(s3_uris)
-    if not ok_uris:
-        return PackageUpdateError(
-            error=uris_error,
-            package_name=package_name,
-            suggested_actions=uris_actions,
-        )
-    ok_package, package_error, package_actions = validate_package_name_required(package_name, "package_update")
-    if not ok_package:
-        return PackageUpdateError(
-            error=package_error,
-            package_name="",
-            suggested_actions=package_actions,
-        )
-    ok_meta, meta_error, _, meta_actions = validate_metadata_dict(metadata, package_name=package_name)
-    if not ok_meta:
-        return PackageUpdateError(
-            error=meta_error,
-            package_name=package_name,
-            suggested_actions=meta_actions,
-        )
-    warnings: list[str] = []
-    normalized_registry = _normalize_registry(registry)
-    auth_ctx, error = _authorize_package(
-        "package_update",
-        {"package_name": package_name, "registry": normalized_registry},
-        context={"package_name": package_name, "registry": normalized_registry},
-    )
-    if error:
-        return PackageUpdateError(
-            error=error.get("error", "Authorization failed"),
-            package_name=package_name,
-            registry=registry,
-            suggested_actions=["Check your permissions", "Verify package exists", "Confirm registry access"],
-        )
-
-    try:
-        # Use QuiltOps.update_package_revision() for business logic
-        from ..ops.factory import QuiltOpsFactory
-        from ..utils.common import suppress_stdout
-
-        quilt_ops = QuiltOpsFactory.create()
-
-        # Convert parameters to match QuiltOps interface
-        copy_behavior = "all" if copy else "none"
-        auto_organize = not flatten  # flatten=True means auto_organize=False
-
-        with suppress_stdout():
-            result = quilt_ops.update_package_revision(
-                package_name=package_name,
-                s3_uris=s3_uris,
-                registry=normalized_registry,
-                metadata=metadata,
-                message=message,
-                auto_organize=auto_organize,
-                copy=copy_behavior,
-            )
-
-        if not result.success:
-            return PackageUpdateError(
-                error="Package update failed - no files were added or push failed",
-                package_name=package_name,
-                registry=registry,
-                suggested_actions=[
-                    "Check that S3 URIs are valid and accessible",
-                    "Verify URIs point to actual files, not directories",
-                    "Ensure you have write permissions to the registry",
-                ],
-                warnings=warnings,
-            )
-
-        # Transform QuiltOps result to tool response format
-        return PackageUpdateSuccess(
-            package_name=package_name,
-            registry=registry,
-            top_hash=str(result.top_hash),  # Convert to string for Pydantic validation
-            files_added=result.file_count,
-            package_url=result.catalog_url or "",
-            files=[],  # QuiltOps doesn't return detailed file list, but that's OK
-            message=message,
-            warnings=warnings,
-            auth_type=auth_ctx.auth_type if auth_ctx else None,
-        )
-
-    except Exception as e:
-        return PackageUpdateError(
-            error=f"Package update failed: {e}",
-            package_name=package_name,
-            registry=registry,
-            suggested_actions=[
-                "Check that the package exists in the registry",
-                "Verify S3 URIs are valid and accessible",
-                "Ensure you have read/write permissions",
-                "Check network connectivity",
-            ],
-            warnings=warnings,
-        )
-
-
-def package_delete(
-    package_name: Annotated[
-        str,
-        Field(
-            description="Name of the package to delete in namespace/name format",
-            examples=["username/dataset", "team/old-analysis"],
-            pattern=r"^[a-zA-Z0-9_-]+/[a-zA-Z0-9_-]+$",
-        ),
-    ],
-    registry: Annotated[
-        str,
-        Field(
-            description="Quilt registry S3 URI where the package resides (REQUIRED)",
-        ),
-    ],
-) -> PackageDeleteSuccess | PackageDeleteError:
-    """Delete a Quilt package from the registry - Core package creation, update, and deletion workflows
-
-    Args:
-        package_name: Name of the package to delete in namespace/name format
-        registry: Quilt registry S3 URI where the package resides (REQUIRED)
-
-    Returns:
-        PackageDeleteSuccess with confirmation, or PackageDeleteError on failure.
-
-    Next step:
-        Report the package operation result or continue the workflow (e.g., metadata updates).
-
-    Example:
-        ```python
-        from quilt_mcp.tools import packages
-
-        result = packages.package_delete(package_name="team/dataset", registry="s3://my-bucket")
-        # Next step: Report the package operation result or continue the workflow (e.g., metadata updates).
-        ```
-    """
-    # Initialize auth_ctx to avoid fragile locals() checks in exception handlers
-    auth_ctx: AuthorizationContext | None = None
-
-    if not package_name:
-        return PackageDeleteError(
-            error="package_name is required for package deletion",
-            package_name="",
-            suggested_actions=["Provide a package name in format: namespace/name", "Example: team/dataset"],
-        )
-
-    # Validate required registry parameter
-    if not registry:
-        return PackageDeleteError(
-            error="Registry parameter is required for package deletion. Specify target S3 bucket (e.g., registry='s3://my-bucket')",
-            package_name=package_name,
-            registry="",
-            suggested_actions=[
-                "Provide registry parameter",
-                "Example: package_delete(package_name='team/dataset', registry='s3://my-bucket')",
-            ],
-        )
-
-    try:
-        normalized_registry = _normalize_registry(registry)
-        auth_ctx, error = _authorize_package(
-            "package_delete",
-            {"package_name": package_name, "registry": normalized_registry},
-            context={"package_name": package_name, "registry": normalized_registry},
-        )
-        if error:
-            return PackageDeleteError(
-                error=error.get("error", "Authorization failed"),
-                package_name=package_name,
-                registry=registry,
-                suggested_actions=["Check your permissions", "Verify package exists", "Confirm registry access"],
-            )
-
-        from ..utils.common import suppress_stdout
-
-        backend = QuiltOpsFactory.create()
-        with suppress_stdout():
-            deleted = backend.delete_package(name=package_name, bucket=normalized_registry)
-
-        if not deleted:
-            return PackageDeleteError(
-                error=f"Failed to delete package '{package_name}'",
-                package_name=package_name,
-                registry=registry,
-                suggested_actions=[
-                    "Verify package exists in the registry",
-                    "Check delete permissions for the registry",
-                    "Ensure package name is correct",
-                ],
-            )
-
-        return PackageDeleteSuccess(
-            package_name=package_name,
-            registry=registry,
-            message=f"Package {package_name} deleted successfully",
-            auth_type=auth_ctx.auth_type if auth_ctx else None,
-        )
-    except Exception as e:
-        return PackageDeleteError(
-            error=f"Failed to delete package '{package_name}': {e}",
-            package_name=package_name,
-            registry=registry,
-            suggested_actions=[
-                "Verify package exists in the registry",
-                "Check delete permissions for the registry",
-                "Ensure package name is correct",
-                "Verify network connectivity",
-            ],
-        )
+# CRUD and browse operations are delegated to package_crud to keep this module focused on S3-ingestion workflows.
+from .package_crud import (
+    package_browse,
+    package_create,
+    package_delete,
+    package_diff,
+    package_update,
+    packages_list,
+)
 
 
 def package_create_from_s3(
